Abstract out the native action registering of DefaultActivityActionHandlers.kt into an intermediate convenience class, MutableListMultiActionHandler.kt
